Introduce entities, types and factories.
I would like the Houses to have random positions until I work on the
backend. This commit introduces types for `House` entities and a factory
with random, but overrideable fields.

I also started to defined some constants and added some comments in the
CSS file so that hopefully the "world size" and the "location size"
match. It's not the most ideal to not have these values connected, but I
can come back to it later if I want to.