Switched to using a `Map` for constraints.
This fixed the bug I was having with adding a removing constraints. It
currently only allows that adding a removing of one preset type of
constraint because the form elements are not functional.

I also added tests for this component. Because of it's very interactive
nature. I am writing the tests with a "TestHelper" pattern that I like
using for things that have a lot of interaction and don't really have
props useful for setting up test cases.