Fixup the broken tests for the `ConstraintForm`.
I used this code base for some brain warm-ups yesterday but ran out of
time when I got stuck on the submit button not propagating events.

This commit introduces a `constraintFactory`. This new factory has some
duplication with the `ConstraintTypes` and `ConstraintOperators` because
they are types and so they don't exist at run time.