Refactor: Lift the state up out of the `Controls`
I am hoping to use a single `useReducer` for most of the state of the
app. This commit lifts just the state of the `Controls` up into the
`App` component and splits the `reducer` and `actions` into their own
files.

My plan is do change the state for the map and the list to be a
`useReducer` rather than a `useState`.