Change the locations to have their own part of app state.
This is a pretty large refactor. It involved changing most of the files
in the app in order to make it work.

It's not quite a straight refactor since it doesn't correctly put
"houses" and "schools" into state when the app loads. My plan for that
is to make a fake client that returns a promise to emulate loading from
an API.

One other thing is that the "hightlighted" element is still using a
`useState` hook rather than the shared application state.