Exclude houses based on the criteria.
This commit starts to filter the houses based on the constraints. It
prevents things from being shown in the `WorldMap` and doesn't show the
items in the `HouseList`.

I sorta wish that the `App` component was tested, but I think that might
be a better place for e2e tests.